refactor: peers patcher
