+++ .changeset/large-turkeys-hope.md
@@ -1,6 +1,6 @@
 ---
-"@pnpm/hooks.read-package-hook": patch
+"@pnpm/hooks.read-package-hook": minor
+"pnpm": minor
 ---
 
-Extends the `pnpm.peerDependencyRules.allowedVersions` package.json option to support the
-`parent>child` selector syntax. This syntax allows for overriding specific peerDependencies.
+Extends the `pnpm.peerDependencyRules.allowedVersions` `package.json` option to support the `parent>child` selector syntax. This syntax allows for extending specific `peerDependencies` [#6108](https://github.com/pnpm/pnpm/pull/6108).
