+++ hooks/read-package-hook/src/index.ts
@@ -1,11 +1,1 @@
-import semver from 'semver'
-
 export { createReadPackageHook } from './createReadPackageHook'
-
-export function isSubRange (superRange: string | undefined, subRange: string) {
-  return !superRange ||
-  subRange === superRange ||
-  semver.validRange(subRange) != null &&
-  semver.validRange(superRange) != null &&
-  semver.subset(subRange, superRange)
-}
