+++ hooks/read-package-hook/src/isSubRange.ts
@@ -1,0 +1,9 @@
+import semver from 'semver'
+
+export function isSubRange (superRange: string | undefined, subRange: string) {
+  return !superRange ||
+  subRange === superRange ||
+  semver.validRange(subRange) != null &&
+  semver.validRange(superRange) != null &&
+  semver.subset(subRange, superRange)
+}
